pyproject.toml
[build-system]
requires = [
    "setuptools>=42",
    "wheel",
]
build-backend = "setuptools.build_meta"

[project]
name = "fmbench"
version = "2.0.27"
description = "Benchmark performance of **any Foundation Model (FM)** deployed on **any AWS Generative AI service**, be it **Amazon SageMaker**, **Amazon Bedrock**, **Amazon EKS**, or **Amazon EC2**. The FMs could be deployed on these platforms either directly through `FMbench`, or, if they are already deployed then also they could be benchmarked through the **Bring your own endpoint** mode supported by `FMBench`."
authors = [
    { name = "Amit Arora", email = "aroraai@amazon.com" },
    { name = "Madhur Prashant", email = "Madhurpt@amazon.com" }
]
license = { file = "LICENSE" }
readme = "README.md"
keywords = ["benchmarking", "sagemaker", "bedrock", "bring your own endpoint", "generative-ai", "foundation-models", "llama3", "ec2", "eks"]
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "requests>=2.32.2",
    "ipywidgets==8.1.1",
    "transformers==4.46.1",
    "pandas==2.1.4",
    "datasets==2.16.1",
    "seaborn==0.13.1",
    "tomark==0.1.4",
    "jupyter>=1.0.0",
    "boto3>=1.34.109",
    "papermill>=2.5.0",
    "pyyaml",
    "sagemaker==2.223.0",
    "litellm==1.54.0",
    "plotly==5.22.0",
    "kaleido==0.2.1",
    "jinja2==3.1.4",
    "tqdm==4.66.4",
    "pydantic==2.7.4",
    "ray==2.32.0",
    "torch==2.4",
    "sentence-transformers==3.0.1",
    "ec2-metadata==2.13.0",
    "nvitop==1.3.2",
    "psutil==5.9.8",
    "pip>=24.3.1",
    "ipykernel>=6.29.5",
    "jupyter-client>=8.6.3",
    "zmq>=0.0.0",
    "vllm>=0.5.4",
]

[tool.setuptools]
packages = ["fmbench"]
include-package-data = true

[project.scripts]
fmbench = "fmbench.main:main"

